+++ include/camera.h
@@ -37,11 +37,11 @@ public:
 
 
     //! @brief  Defintion of a position type for convenience
-    typedef std::tuple<double, double, double> Position;
+    typedef std::tuple<float, float, float> Position;
 
 
     //! @brief  Defintion of a poiting vector type for convenience
-    typedef std::tuple<double, double, double> Pointing_Vector;
+    typedef std::tuple<float, float, float> Pointing_Vector;
 
 
     //! @brief Constructor for a model of a camera
@@ -49,7 +49,7 @@ public:
     //! @param[in] field_of_view_rad    The field of view of the camera, in radians
     //! @param[in] rows                 The size of the vertical focal plane dimension, in pixels
     //! @param[in] cols                 The size of the horizontal focal plane dimension, in pixels
-    Camera(const double field_of_view_rad,
+    Camera(const float field_of_view_rad,
            const uint16_t rows, 
            const uint16_t cols);
 
@@ -67,17 +67,17 @@ public:
 
 
     //! @brief  Retrieve the focal length of the Camera, in pixels based on intrinsic parameters
-    double focal_length() const;
+    float focal_length() const;
 
 
     //! @brief  Retreieve the field of view of the sensor in radians
-    double fov() const;
+    float fov() const;
 
 
     //! @brief Set the field of view of the Camera
     //!
     //! @param[in]  fov_rad     The new field of view of the Camera, in radians
-    void set_fov(const double fov_rad);
+    void set_fov(const float fov_rad);
 
 
     //! @brief  Retrieve the focal plane dimensions of the camera, in pixels
@@ -106,7 +106,7 @@ public:
 private:
 
     //! The field of view of the Camera, in radians
-    double m_field_of_view_rad;
+    float m_field_of_view_rad;
 
     //! The dimensions of the focal plane, in pixels
     Size m_focal_plane_size_pixels;
