+++ src/diamond_square_terrain_generator.cc
@@ -34,8 +34,8 @@ Diamond_Square_Generator::~Diamond_Square_Generator()
 
 Terrain Diamond_Square_Generator::generate_terrain(const uint32_t rows, 
                                                    const uint32_t cols, 
-                                                   const double scale, 
-                                                   const double roughness)
+                                                   const float scale, 
+                                                   const float roughness)
 {
     if (rows != cols) {
         throw std::invalid_argument("Diamond-Square Generator requires that rows == cols");
@@ -58,17 +58,17 @@ Terrain Diamond_Square_Generator::generate_terrain(const uint32_t rows,
     std::mt19937 gen(rd());
 
     while (half >= 1) {
-        const double feature_scale = size * roughness;
+        const float feature_scale = size * roughness;
 
         auto rng = std::uniform_real_distribution<>(-feature_scale, feature_scale);
 
         // Process squares
         for (uint32_t r = half; r < rows; r += size) {
             for (uint32_t c = half; c < cols; c += size) {
-                const double offset = rng(gen);
+                const float offset = rng(gen);
                 
-                const double sum = tbuffer.at(r - half, c - half) + tbuffer.at(r - half, c + half) +
-                                   tbuffer.at(r + half, c - half) + tbuffer.at(r + half, c + half);
+                const float sum = tbuffer.at(r - half, c - half) + tbuffer.at(r - half, c + half) +
+                                  tbuffer.at(r + half, c - half) + tbuffer.at(r + half, c + half);
 
                 tbuffer.at(r, c) = (sum / 4) + offset;
             }
@@ -77,10 +77,10 @@ Terrain Diamond_Square_Generator::generate_terrain(const uint32_t rows,
         // Process diamonds
         for (uint32_t r = 0; r < rows; r += half) {
             for (uint32_t c = 0; c < cols; c += half) {
-                const double offset = rng(gen);
+                const float offset = rng(gen);
                 
-                const double sum = tbuffer.at(r, c - size) + tbuffer.at(r + size, c) + 
-                                   tbuffer.at(r, c + size) + tbuffer.at(r - size, c);
+                const float sum = tbuffer.at(r, c - size) + tbuffer.at(r + size, c) + 
+                                  tbuffer.at(r, c + size) + tbuffer.at(r - size, c);
 
                 tbuffer.at(r, c) = (sum / 4) + offset;
             }
